PATCH 0: the presentation file managed by the application
The presentation page ("/services/presentation.html" by default and "/web/ushare.html" for ushare server) is managed by the application and not byt the library. It's the application which set the file (name and contain) and  checks the requests about it (check the HTTP document request and look for the filename).

The default directory haven't to be hard define and should be modify by the application.