dlna_item_new becomes function API
to be able to manage the VFS and its contain, the user application has
to know the item that it inserts into the VFS. To do that dlna_item_new
create an item from the file path. After the application user can insert
it into the wishy container (or containers).