create stream seekable to check files properties and play stream
dlna_stream_t new structure with read, lseek callback.
This structure is create by the libdlna and passed to the profiler.
The profiler has to use the callbacks to parse the files and checks the contain.
When all profilers used the stream, libdlna close this one before to check another stream.